I wanted to make generic the shaping idea
and figure out how to make it run in parallel

Then I took a look at the current cake code which
got really complex compared to what I remember.

my idea for basic parallelism is to do an atomic_add
on a static variable for time_next_packet, and figure
out how it could be sanely shared between invocations
later.